requirements.txt
streamlit 
pymupdf 
pillow 
google-generativeai
 groq 
 python-dotenv 
 camelot-py[pdf] 
 sentence-transformers
  faiss-cpu 
  numpy